+++ CHANGELOG.md
@@ -8,6 +8,19 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 0.22.2 - 2021-04-26
+
+### Changed
+
+- Raise `project/{PROJECT_ID} is not a workspace` error message when attempting
+  to fetch alert monitoring policies
+
+### Fixed
+
+- Fixed [#107](https://github.com/JupiterOne/graph-google-cloud/issues/107) -
+  Publish unprocessed buckets message to integration job log when a bucket is
+  configured with "requestor pays"
+
 ## 0.22.1 - 2021-04-22
 
 ### Changed